pyproject.toml
[project]
name = "menu-analyzer-ai"
version = "0.1.0"
description = "AI-powered menu analysis and recommendation system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.26.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-openai>=0.3.14",
    "pillow>=11.2.1",
    "ruff>=0.11.7",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pytest>=8.3.5",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "requests>=2.32.3",
]
